+++ TBMac/TBViewerViewController.m
@@ -136,7 +136,7 @@
 - (NSView *)tableView:(NSTableView*)aTableView viewForTableColumn:(NSTableColumn*)aTableColumn row:(NSInteger)rowIndex {
     NSTableCellView* result = [aTableView makeViewWithIdentifier:aTableColumn.identifier owner:self];
     if([[aTableColumn identifier] isEqualToString:@"Payout"]) {
-        NSDictionary* object = [[[self resultsArrayController] arrangedObjects] objectAtIndex:rowIndex];
+        NSDictionary* object = [[[self resultsArrayController] arrangedObjects] objectAtIndex:(NSUInteger)rowIndex];
         [[[result textField] formatter] setCurrencyCode:object[@"payout"][@"currency"]];
     } else if([[aTableColumn identifier] isEqualToString:@"Chip"]) {
         TBChipTableCellView* chipTableCellView = (TBChipTableCellView*)result;
